refactor: Update dcdio_module.c to use new C implementation
This commit updates the dcdio_module.c file in the dcdio extension to use the new C implementation for the open_dcd_read function. The new implementation improves performance and removes the dependency on the SWIG library.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (refactor for code refactoring).