Passed integration test for test_intg_file_io_Files_close_dcd_read.py
Updated ./clean command for local testing.

Add check_modules.py to list all modules in a package

This commit adds a new file, check_modules.py, to the dcdio extension. The file contains a function, list_modules, that lists all the modules in a given package using the pkgutil module. This functionality will be useful for future development and debugging.